Move urql context into just search page
